fix(rebase): resolve cli.rs refactor conflicts; re-apply AI incidents/investigate/assess
Main extracted CliMode/GlobalFlags/run() to cli/run.rs after this branch diverged.
Re-apply ai-abuse additions to the correct split files:
- cli/args.rs: add Incidents/Investigate/Assess variants + AiIncidentsArgs/AiInvestigateArgs/AiAssessArgs
- cli/run.rs: add dispatch arms for the 3 new AiCommand variants
- cli.rs: add parse_ai_incidents/investigate/assess fns + print fns + response imports
