+++ src/cli/args.rs
@@ -38,9 +38,9 @@ pub(crate) enum AiCommand {
     Doctor(AiDoctorArgs),
     WatchStatus(OutputArgs),
     SmokeWatch(OutputArgs),
-    SimilarIncidents(AiSimilarArgs),
-    AskHistory(AiAskHistoryArgs),
-    IncidentContext(AiIncidentContextArgs),
+    Incidents(AiIncidentsArgs),
+    Investigate(AiInvestigateArgs),
+    Assess(AiAssessArgs),
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -394,39 +394,61 @@ pub(crate) struct AiPruneCheckpointsArgs {
 }
 
 #[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub(crate) struct AiSimilarArgs {
-    pub query: String,
-    pub hostname: Option<String>,
-    pub app_name: Option<String>,
-    pub severity_min: Option<String>,
+pub(crate) struct AiIncidentsArgs {
+    pub project: Option<String>,
+    pub tool: Option<String>,
+    pub from: Option<String>,
+    pub to: Option<String>,
+    pub limit: Option<u32>,
+    pub window_minutes: Option<u32>,
+    pub terms: Vec<String>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, Default, PartialEq, Eq)]
+pub(crate) struct AiInvestigateArgs {
+    pub project: Option<String>,
+    pub tool: Option<String>,
+    pub from: Option<String>,
+    pub to: Option<String>,
+    pub limit: Option<u32>,
+    pub window_minutes: Option<u32>,
+    pub correlation_window_minutes: Option<u32>,
+    pub terms: Vec<String>,
+    pub json: bool,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct AiAssessArgs {
+    pub incident_id: String,
+    pub model: Option<String>,
+    pub json: bool,
+    pub project: Option<String>,
+    pub tool: Option<String>,
     pub from: Option<String>,
     pub to: Option<String>,
     pub window_minutes: Option<u32>,
+    pub correlation_window_minutes: Option<u32>,
+    pub terms: Vec<String>,
     pub limit: Option<u32>,
-    pub json: bool,
 }
 
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub(crate) struct AiAskHistoryArgs {
-    pub query: String,
-    pub hostname: Option<String>,
-    pub app_name: Option<String>,
-    pub from: Option<String>,
-    pub to: Option<String>,
-    pub limit: Option<u32>,
-    pub json: bool,
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub(crate) struct AiIncidentContextArgs {
-    pub from: String,
-    pub to: String,
-    pub hostname: Option<String>,
-    pub app_name: Option<String>,
-    pub query: Option<String>,
-    pub severity_min: Option<String>,
-    pub limit: Option<u32>,
-    pub json: bool,
+impl Default for AiAssessArgs {
+    fn default() -> Self {
+        Self {
+            incident_id: String::new(),
+            model: None,
+            json: false,
+            project: None,
+            tool: None,
+            from: None,
+            to: None,
+            window_minutes: None,
+            correlation_window_minutes: None,
+            terms: Vec::new(),
+            limit: None,
+        }
+    }
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
