+++ src/cli/run.rs
@@ -72,13 +72,9 @@ pub(crate) async fn run(mode: CliMode, command: CliCommand) -> Result<()> {
             AiCommand::SmokeWatch(args) => dispatch::run_ai_smoke_watch(&mode, args).await,
             AiCommand::WatchStatus(args) => dispatch::run_ai_watch_status(&mode, args).await,
             AiCommand::Watch(args) => dispatch::run_ai_watch(&mode, args).await,
-            AiCommand::SimilarIncidents(args) => {
-                dispatch::run_ai_similar_incidents(&mode, args).await
-            }
-            AiCommand::AskHistory(args) => dispatch::run_ai_ask_history(&mode, args).await,
-            AiCommand::IncidentContext(args) => {
-                dispatch::run_ai_incident_context(&mode, args).await
-            }
+            AiCommand::Incidents(args) => dispatch::run_ai_incidents(&mode, args).await,
+            AiCommand::Investigate(args) => dispatch::run_ai_investigate(&mode, args).await,
+            AiCommand::Assess(args) => dispatch::run_ai_assess(&mode, args).await,
         },
         // DB commands (bead 0p8r.9). 4 are HTTP-capable; backup stays LOCAL
         // and bails in HTTP mode with an inline message.
